// helper/types.ts
export interface State {
    theme: {
        modal: boolean,
        fontSize: {
            xs: string;
            sm: string;
            base: string;
            lg: string;
            xl: string;
            xxl: string;
            xxxl: string;
            xxxxl: string;
            xxxxxl: string;
            xxxxxxl: string;
        },
        color: string;
        background: {
            secondary: string;
            primary: string;
            textprimary: string;
            textsecondary: string;
        },
    }
}